Wifi script for Ubuntu
In case the wifi doesnt work properly or is connected but no speed in
your wifi, just run this script.

Note:
This works in ubuntu 14.04 and chmod the file if you wanna run
it as a script and add it to your sys path if you wanna call it
without its path and just filename.

1) PATH=/home/<path_to_file>/wifi_ubuntu_script.sh:"$PATH"
2) chmod +x wifi_ubuntu_script.sh